bumps to 1.0.0-beta
